+++ AridArnold/Arcade/HorsesAndGun/HorsesAndGun.cs
@@ -65,7 +65,7 @@
 			return screenTargetRef;
 		}
 
-		public override int GetScore()
+		public override ulong GetScore()
 		{
 			return ScoreManager.I.GetCurrentScore();
 		}
